Update footer to be a separate component. Change footer styling.
